Required fixes to build
